requirements.txt
pymupdf==1.22.3
pytesseract==0.3.10
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.335
chromadb==0.4.18
pydantic==2.4.2
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6 